refactor: create sensor table cell compontent for overview exposure table
